feat(middleware): move verifyRequestBody file
